:bug: 途中でphiloが停止する問題を解決

+++ main.c
@@ -6,7 +6,7 @@
 /*   By: yooshima <yooshima@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/26 13:08:52 by yooshima          #+#    #+#             */
-/*   Updated: 2024/08/28 13:50:38 by yooshima         ###   ########.fr       */
+/*   Updated: 2024/08/31 11:25:10 by yooshima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,14 +36,11 @@ int	main(int argc, char *argv[])
 	{
 		return (write(2, "Error:Invalid args. 3 700 200 200 [10]\n", 39), 1);
 	}
-	// if (fork_init(ft_atoi(argv[1]), fork) == -1 \
-	// 	|| philo_init(argc, argv, philo, fork) == -1)
-	// 	return (1);
-	fork_init(ft_atoi(argv[1]), fork);
-	philo_init(argc, argv, philo, fork);
-	// if (thread_make(philo, fork) != 0)
-	// 	return (destroy_all(philo, fork), 1);
-	thread_make(philo, fork);
+	if (fork_init(ft_atoi(argv[1]), fork) == -1 \
+		|| philo_init(argc, argv, philo, fork) == -1)
+		return (1);
+	if (thread_make(philo, fork) != 0)
+		return (destroy_all(philo, fork), 1);
 	destroy_all(philo, fork);
 	return (0);
 }
